requirements.txt
discord.py
firebase-admin
python-dotenv
google-generativeai==0.8.3
ruff
black
isort
mypy